+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 ### Fix
 * Throw error if ride has the same departure and arrival POI
+* Added rate limit model & handles it for password attempt
 ## Version 1.4.5
 ### Features
 * Added connection status management for drivers through sockets